[REFACTOR] - Refactor in pkgs
