+++ app/src/main/java/com/example/login_compose/Login.kt
@@ -1,40 +1,16 @@
 package com.example.login_compose
 
-import android.app.Activity
-import androidx.compose.foundation.Image
-import androidx.compose.foundation.background
-import androidx.compose.foundation.clickable
-import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.Close
-import androidx.compose.material3.Button
-import androidx.compose.material3.Divider
-import androidx.compose.material3.Icon
-import androidx.compose.material3.Text
-import androidx.compose.material3.TextField
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.saveable.rememberSaveable
-import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.platform.LocalContext
-import androidx.compose.ui.res.painterResource
-import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import androidx.compose.ui.unit.sp
+import com.example.login_compose.components.login.BodyLogin
+import com.example.login_compose.components.login.FooterLogin
+import com.example.login_compose.components.login.HeaderLogin
 
 
 @Composable
@@ -45,176 +21,11 @@ fun Login() {
             .padding(20.dp)
     ) {
         HeaderLogin(Modifier.align(Alignment.TopEnd))
-        Body(Modifier.align(Alignment.Center))
-        Footer(Modifier.align(Alignment.BottomCenter))
+        BodyLogin(Modifier.align(Alignment.Center))
+        FooterLogin(Modifier.align(Alignment.BottomCenter))
     }
 
 }
-
-@Composable
-fun Footer(modifier: Modifier) {
-    Column(modifier = modifier.fillMaxWidth()) {
-        Divider(
-            Modifier
-                .background(Color(0xFFF9F9F9))
-                .height(1.dp)
-                .fillMaxWidth()
-        )
-        Spacer(modifier = Modifier.size(16.dp))
-        SingUp()
-        Spacer(modifier = Modifier.size(16.dp))
-
-    }
-}
-
-@Composable
-fun SingUp() {
-    Row(Modifier.fillMaxWidth(), horizontalArrangement = Arrangement.Center) {
-        Text(text = "Don`t have an account?", fontSize = 12.sp, color = Color(0xFFB5B5B5))
-        Text(
-            text = "Sing up.",
-            Modifier.padding(horizontal = 8.dp),
-            fontSize = 12.sp,
-            fontWeight = FontWeight.Bold,
-            color = Color(0xFF4EA8E9)
-        )
-    }
-}
-
-@Composable
-fun Body(modifier: Modifier) {
-    var email by rememberSaveable {
-        mutableStateOf("")
-    }
-    var password by rememberSaveable {
-        mutableStateOf("")
-    }
-    var isLoginEnable by rememberSaveable {
-        mutableStateOf(false)
-    }
-
-    Column(modifier = modifier) {
-        ImageLogo(modifier.align(Alignment.CenterHorizontally))
-        Spacer(modifier = Modifier.size(16.dp))
-        Email(email = email) { email = it }
-        Spacer(modifier = Modifier.size(4.dp))
-        Password(password = password) { password = it }
-        Spacer(modifier = Modifier.size(8.dp))
-        ForgotPassword(Modifier.align(Alignment.End))
-        Spacer(modifier = Modifier.size(16.dp))
-        LoginButton(isLoginEnable)
-        Spacer(modifier = Modifier.size(16.dp))
-        LoginDivider()
-        Spacer(modifier = Modifier.size(32.dp))
-        SocialLogin()
-    }
-}
-
-@Composable
-fun SocialLogin() {
-    Row(
-        Modifier.fillMaxWidth(),
-        verticalAlignment = Alignment.CenterVertically,
-        horizontalArrangement = Arrangement.Center
-    ) {
-        Image(
-            painter = painterResource(id = R.drawable.fb),
-            contentDescription = "Social login fb",
-            modifier = Modifier.size(19.dp)
-        )
-        Text(
-            text = "Continue as Martin",
-            fontSize = 14.sp,
-            fontWeight = FontWeight.Bold,
-            modifier = Modifier.padding(8.dp),
-            color = Color(0XFF4EA8E9)
-        )
-    }
-}
-
-@Composable
-fun LoginDivider() {
-    Row(Modifier.fillMaxWidth(), verticalAlignment = Alignment.CenterVertically) {
-        Divider(
-            Modifier
-                .background(Color(0xFFF9F9F9))
-                .height(1.dp)
-                .weight(1f)
-        )
-        Text(
-            text = "OR",
-            modifier = Modifier.padding(horizontal = 18.dp),
-            fontSize = 12.sp,
-            fontWeight = FontWeight.Bold,
-            color = Color(0xFFB5B5B5)
-        )
-        Divider(
-            Modifier
-                .background(Color(0xFFF9F9F9))
-                .height(1.dp)
-                .weight(1f)
-        )
-    }
-}
-
-@Composable
-fun LoginButton(loginEnable: Boolean) {
-    Button(onClick = { /*TODO*/ }, enabled = loginEnable, modifier = Modifier.fillMaxWidth()) {
-        Text(text = "Log In")
-    }
-}
-
-@Composable
-fun ImageLogo(modifier: Modifier) {
-    Image(
-        painter = painterResource(id = R.drawable.insta),
-        contentDescription = "logo",
-        modifier = modifier
-    )
-}
-
-@Composable
-fun Email(email: String, onTextChange: (String) -> Unit) {
-
-    TextField(
-        value = email,
-        onValueChange = { onTextChange(it) },
-        modifier = Modifier.fillMaxWidth()
-    )
-}
-
-@Composable
-fun Password(password: String, onTextChange: (String) -> Unit) {
-
-    TextField(
-        value = password,
-        onValueChange = { onTextChange(it) },
-        modifier = Modifier.fillMaxWidth()
-    )
-}
-
-@Composable
-fun ForgotPassword(modifier: Modifier) {
-    Text(
-        text = "Forgot password?",
-        fontSize = 12.sp,
-        fontWeight = FontWeight.Bold,
-        color = Color.Blue,
-        modifier = modifier
-    )
-}
-
-
-@Composable
-fun HeaderLogin(modifier: Modifier) {
-    val activity = LocalContext.current as Activity
-    Icon(
-        imageVector = Icons.Default.Close,
-        contentDescription = "close app",
-        modifier = modifier.clickable { activity.finish() })
-
-}
-
 
 @Preview(showBackground = true)
 @Composable
