+++ app/src/main/java/com/example/login_compose/components/login/BodyLogin.kt
@@ -1,0 +1,153 @@
+package com.example.login_compose.components.login
+
+import androidx.compose.foundation.Image
+import androidx.compose.foundation.background
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
+import androidx.compose.material3.Button
+import androidx.compose.material3.Divider
+import androidx.compose.material3.Text
+import androidx.compose.material3.TextField
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.saveable.rememberSaveable
+import androidx.compose.runtime.setValue
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+import com.example.login_compose.R
+
+@Composable
+fun BodyLogin(modifier: Modifier) {
+    var email by rememberSaveable {
+        mutableStateOf("")
+    }
+    var password by rememberSaveable {
+        mutableStateOf("")
+    }
+    var isLoginEnable by rememberSaveable {
+        mutableStateOf(false)
+    }
+
+    Column(modifier = modifier) {
+        ImageLogo(modifier.align(Alignment.CenterHorizontally))
+        Spacer(modifier = Modifier.size(16.dp))
+        Email(email = email) { email = it }
+        Spacer(modifier = Modifier.size(4.dp))
+        Password(password = password) { password = it }
+        Spacer(modifier = Modifier.size(8.dp))
+        ForgotPassword(Modifier.align(Alignment.End))
+        Spacer(modifier = Modifier.size(16.dp))
+        LoginButton(isLoginEnable)
+        Spacer(modifier = Modifier.size(16.dp))
+        LoginDivider()
+        Spacer(modifier = Modifier.size(32.dp))
+        SocialLogin()
+    }
+}
+
+@Composable
+fun ImageLogo(modifier: Modifier) {
+    Image(
+        painter = painterResource(id = R.drawable.insta),
+        contentDescription = "logo",
+        modifier = modifier
+    )
+}
+
+@Composable
+fun Email(email: String, onTextChange: (String) -> Unit) {
+
+    TextField(
+        value = email,
+        onValueChange = { onTextChange(it) },
+        modifier = Modifier.fillMaxWidth()
+    )
+}
+
+@Composable
+fun Password(password: String, onTextChange: (String) -> Unit) {
+
+    TextField(
+        value = password,
+        onValueChange = { onTextChange(it) },
+        modifier = Modifier.fillMaxWidth()
+    )
+}
+
+
+@Composable
+fun ForgotPassword(modifier: Modifier) {
+    Text(
+        text = "Forgot password?",
+        fontSize = 12.sp,
+        fontWeight = FontWeight.Bold,
+        color = Color.Blue,
+        modifier = modifier
+    )
+}
+
+@Composable
+fun LoginButton(loginEnable: Boolean) {
+    Button(onClick = { /*TODO*/ }, enabled = loginEnable, modifier = Modifier.fillMaxWidth()) {
+        Text(text = "Log In")
+    }
+}
+
+@Composable
+fun LoginDivider() {
+    Row(Modifier.fillMaxWidth(), verticalAlignment = Alignment.CenterVertically) {
+        Divider(
+            Modifier
+                .background(Color(0xFFF9F9F9))
+                .height(1.dp)
+                .weight(1f)
+        )
+        Text(
+            text = "OR",
+            modifier = Modifier.padding(horizontal = 18.dp),
+            fontSize = 12.sp,
+            fontWeight = FontWeight.Bold,
+            color = Color(0xFFB5B5B5)
+        )
+        Divider(
+            Modifier
+                .background(Color(0xFFF9F9F9))
+                .height(1.dp)
+                .weight(1f)
+        )
+    }
+}
+
+@Composable
+fun SocialLogin() {
+    Row(
+        Modifier.fillMaxWidth(),
+        verticalAlignment = Alignment.CenterVertically,
+        horizontalArrangement = Arrangement.Center
+    ) {
+        Image(
+            painter = painterResource(id = R.drawable.fb),
+            contentDescription = "Social login fb",
+            modifier = Modifier.size(19.dp)
+        )
+        Text(
+            text = "Continue as Martin",
+            fontSize = 14.sp,
+            fontWeight = FontWeight.Bold,
+            modifier = Modifier.padding(8.dp),
+            color = Color(0XFF4EA8E9)
+        )
+    }
+}
