+++ app/src/main/java/com/example/login_compose/components/login/HeaderLogin.kt
@@ -1,0 +1,20 @@
+package com.example.login_compose.components.login
+
+import android.app.Activity
+import androidx.compose.foundation.clickable
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.Close
+import androidx.compose.material3.Icon
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.platform.LocalContext
+
+@Composable
+fun HeaderLogin(modifier: Modifier) {
+    val activity = LocalContext.current as Activity
+    Icon(
+        imageVector = Icons.Default.Close,
+        contentDescription = "close app",
+        modifier = modifier.clickable { activity.finish() })
+
+}
